requirements-test.txt
# Testing requirements for Energy Generation Dashboard

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0

# Test utilities
pytest-html>=3.1.0
pytest-json-report>=1.5.0
pytest-benchmark>=4.0.0
pytest-asyncio>=0.21.0

# Mocking and fixtures
responses>=0.23.0
factory-boy>=3.2.0
freezegun>=1.2.0

# Data testing
pandas-testing>=0.0.1
numpy-testing>=1.0.0

# Coverage reporting
coverage>=7.0.0
coverage-badge>=1.1.0

# Code quality (for testing)
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Performance testing
memory-profiler>=0.60.0
psutil>=5.9.0

# Test data generation
Faker>=18.0.0
hypothesis>=6.70.0

# Streamlit testing utilities
streamlit>=1.28.0

# Additional utilities for testing
tox>=4.0.0
pre-commit>=3.0.0
